Employee Add changes. EmployeeHourlyRateDto changes.

+++ Sourcecode/API/Strive.BusinessEntities/DTO/Employee/EmployeeHourlyRateDto.cs
@@ -6,8 +6,7 @@
         public int EmployeeId { get; set; }
         public int LocationId { get; set; }
         public int? RoleId { get; set; }
-
-        public int HourlyRate { get; set; }
+        public decimal HourlyRate { get; set; }
         public string RoleName { get; set; }
         public string LocationName { get; set; }
     }
